+++ CHANGES.rst
@@ -6,6 +6,8 @@
   used on :class:`udev.Device` objects.
 - added :meth:`udev.Enumerator.match_children` and
   :attr:`udev.Device.children` to list direct children of a device
+- added :pyqt4:sig:`qudev.QUDevMonitorObserver.deviceChanged` and
+  :pyqt4:sig:`qudev.QUDevMonitorObserver.deviceMoved`
 
 
 0.2 (Jun 28, 2010)